# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the Manager Dashboard application. Provides endpoints for user management, team organization, time-off requests, Jira integration, and calendar management.",
        "title": "Manager Dashboard API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/invitations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all invitations with pagination support. Admin only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitations"
                ],
                "summary": "Get all invitations",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of invitations",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_smith-dallin_manager-dashboard_internal_models.Invitation"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden - admin access required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new invitation to onboard a user. Admin only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitations"
                ],
                "summary": "Create a new invitation",
                "parameters": [
                    {
                        "description": "Invitation request",
                        "name": "invitation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_smith-dallin_manager-dashboard_internal_models.CreateInvitationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created invitation",
                        "schema": {
                            "$ref": "#/definitions/github_com_smith-dallin_manager-dashboard_internal_models.Invitation"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden - admin access required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "User or invitation already exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/invitations/accept/{token}": {
            "post": {
                "description": "Accepts an invitation and creates the user account. Called after Auth0 signup.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitations"
                ],
                "summary": "Accept an invitation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invitation token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "auth0_id": {
                                    "type": "string"
                                },
                                "first_name": {
                                    "type": "string"
                                },
                                "last_name": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created user",
                        "schema": {
                            "$ref": "#/definitions/github_com_smith-dallin_manager-dashboard_internal_models.User"
                        }
                    },
                    "400": {
                        "description": "Invalid request or expired token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/invitations/validate/{token}": {
            "get": {
                "description": "Validates an invitation token and returns invitation details. Public endpoint.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitations"
                ],
                "summary": "Validate an invitation token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invitation token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invitation validation result",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Token is required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Invalid invitation token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/invitations/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single invitation by ID. Admin only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitations"
                ],
                "summary": "Get invitation by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invitation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Invitation details",
                        "schema": {
                            "$ref": "#/definitions/github_com_smith-dallin_manager-dashboard_internal_models.Invitation"
                        }
                    },
                    "400": {
                        "description": "Invalid invitation ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden - admin access required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Invitation not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revokes a pending invitation. Admin only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invitations"
                ],
                "summary": "Revoke an invitation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Invitation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Invitation revoked"
                    },
                    "400": {
                        "description": "Invalid invitation ID or already processed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden - admin access required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_smith-dallin_manager-dashboard_internal_models.CreateInvitationRequest": {
            "type": "object",
            "properties": {
                "department": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_smith-dallin_manager-dashboard_internal_models.Role"
                },
                "squad_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "github_com_smith-dallin_manager-dashboard_internal_models.Invitation": {
            "type": "object",
            "properties": {
                "accepted_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "invited_by": {
                    "$ref": "#/definitions/github_com_smith-dallin_manager-dashboard_internal_models.User"
                },
                "invited_by_id": {
                    "type": "integer"
                },
                "role": {
                    "$ref": "#/definitions/github_com_smith-dallin_manager-dashboard_internal_models.Role"
                },
                "squad_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "status": {
                    "$ref": "#/definitions/github_com_smith-dallin_manager-dashboard_internal_models.InvitationStatus"
                },
                "token": {
                    "description": "Only shown to admin who created it",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_smith-dallin_manager-dashboard_internal_models.InvitationStatus": {
            "type": "string",
            "enum": [
                "pending",
                "accepted",
                "expired",
                "revoked"
            ],
            "x-enum-varnames": [
                "InvitationStatusPending",
                "InvitationStatusAccepted",
                "InvitationStatusExpired",
                "InvitationStatusRevoked"
            ]
        },
        "github_com_smith-dallin_manager-dashboard_internal_models.Role": {
            "type": "string",
            "enum": [
                "admin",
                "supervisor",
                "employee"
            ],
            "x-enum-varnames": [
                "RoleAdmin",
                "RoleSupervisor",
                "RoleEmployee"
            ]
        },
        "github_com_smith-dallin_manager-dashboard_internal_models.Squad": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_smith-dallin_manager-dashboard_internal_models.User": {
            "type": "object",
            "properties": {
                "auth0_id": {
                    "type": "string"
                },
                "avatar_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date_started": {
                    "type": "string"
                },
                "department": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "jira_account_id": {
                    "description": "Jira account matching (for org-wide Jira connection)",
                    "type": "string"
                },
                "jira_cloud_id": {
                    "type": "string"
                },
                "jira_domain": {
                    "description": "Jira integration fields (legacy API token auth)",
                    "type": "string"
                },
                "jira_email": {
                    "type": "string"
                },
                "jira_site_url": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_smith-dallin_manager-dashboard_internal_models.Role"
                },
                "squads": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_smith-dallin_manager-dashboard_internal_models.Squad"
                    }
                },
                "supervisor_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}